tx.report sampling review round: 64-bit cadence counters, ceil expectation, aggregation.md caught up
- The 16-bit tag counter's wrap at 65,536 jumps the sampling phase for any N
  that doesn't divide it — off-modulo tag deltas masquerading as dropped
  reports at every seam; J1's 32-bit request counter hits the same seam in
  ~20 days at field rates. Both are 64-bit now (seam unreachable); the tag
  byte and checksummed-span handling are unchanged.
- The sampled expectation ceils (the request fires on k%N==0, frame 0
  first) — odd totals no longer undercount by one.
- docs/aggregation.md's tx.report paragraph carries the divisor semantics
  and the ceiling instead of "every descriptor".

Validated: 48/48 ctest; on-air N=2 smoke under the widened counters —
sampled-coverage 100.0% (82,465/82,490 with the ceil expectation), zero
lost, zero off-modulo anomalies.

+++ src/jaguar1/RtlJaguarDevice.cpp
@@ -1171,7 +1171,7 @@ size_t RtlJaguarDevice::build_tx_block(const uint8_t *packet, size_t length,
   if (_cfg.tx.report)
     SET_TX_DESC_SPE_RPT_8812(
         usb_frame,
-        _tx_ccx_ctr.fetch_add(1) % static_cast<uint32_t>(_cfg.tx.report) == 0
+        _tx_ccx_ctr.fetch_add(1) % static_cast<uint64_t>(_cfg.tx.report) == 0
             ? 1
             : 0);
   SET_TX_DESC_RETRY_LIMIT_ENABLE_8812(usb_frame, 1);
