+++ src/jaguar3/RtlJaguar3Device.cpp
@@ -2006,10 +2006,10 @@ size_t RtlJaguar3Device::build_tx_block(const uint8_t *packet, size_t length,
      * emission saturates at ~1.3k reports/s, and with the tag continuous a
      * received-report tag delta other than N is a dropped report. Both
      * fields sit inside the checksummed span — re-checksum (idempotent). */
-    const uint16_t k = _tx_rpt_tag.fetch_add(1);
+    const uint64_t k = _tx_rpt_tag.fetch_add(1);
     SET_TX_DESC_SPE_RPT_8822C(
-        out, k % static_cast<uint16_t>(_cfg.tx.report) == 0 ? 1 : 0);
-    SET_TX_DESC_SW_DEFINE_8822C(out, k & 0xff);
+        out, k % static_cast<uint64_t>(_cfg.tx.report) == 0 ? 1 : 0);
+    SET_TX_DESC_SW_DEFINE_8822C(out, static_cast<uint16_t>(k & 0xff));
     jaguar3::cal_txdesc_chksum_8822c(out);
   }
   /* Per-frame retry limit from cfg (DEVOURER_TX_RETRY_LIMIT, default 0) —
